requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
scikit-learn==1.6.1
torch==2.6.0
pandas==2.2.3
numpy==2.2.4
supabase==2.13.0
python-dotenv==1.1.0
shap>=0.45.0
joblib==1.4.2
python-multipart==0.0.20
requests==2.32.3
